+++ packages/fiori-app-sub-generator/src/types/external.ts
@@ -229,7 +229,6 @@ export interface ReadonlyState {
 }
 
 export const defaultPromptValues = {
-    [ui5AppInquirerPromptNames.enableCodeAssist]: false,
     [ui5AppInquirerPromptNames.enableEslint]: false,
     [ui5AppInquirerPromptNames.skipAnnotations]: false,
     [ui5AppInquirerPromptNames.enableTypeScript]: false