+++ packages/odata-service-inquirer/src/prompts/datasources/cap-project/cap-helpers.ts
@@ -14,7 +14,7 @@ import LoggerHelper from '../../logger-helper';
 import { errorHandler } from '../../prompt-helpers';
 import type { CapProjectChoice, CapProjectPaths, CapProjectRootPath } from './types';
 import { ERROR_TYPE } from '@sap-ux/inquirer-common';
-import { realpath } from 'fs/promises';
+import { realpath } from 'node:fs/promises';
 
 export const enterCapPathChoiceValue = 'enterCapPath';