+++ packages/ui5-info/src/commands.ts
@@ -1,4 +1,4 @@
-import { spawn } from 'child_process';
+import { spawn } from 'node:child_process';
 
 /**
  * Executes an npm command to extract the supported npm versions from the npm registry.