+++ packages/control-property-editor/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/control-property-editor
 
+## 0.7.6
+
+### Patch Changes
+
+- a9471d0: fix sonar issues
+
 ## 0.7.5
 
 ### Patch Changes