+++ packages/ui5-library-reference-inquirer/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @sap-ux/ui5-library-reference-inquirer
 
+## 0.4.88
+
+### Patch Changes
+
+- Updated dependencies [6382440]
+    - @sap-ux/inquirer-common@0.10.6
+
+## 0.4.87
+
+### Patch Changes
+
+- @sap-ux/inquirer-common@0.10.5
+
+## 0.4.86
+
+### Patch Changes
+
+- @sap-ux/inquirer-common@0.10.4
+
+## 0.4.85
+
+### Patch Changes
+
+- Updated dependencies [c7f9a60]
+    - @sap-ux/project-access@1.33.0
+    - @sap-ux/inquirer-common@0.10.3
+
+## 0.4.84
+
+### Patch Changes
+
+- a9471d0: fix sonar issues
+- Updated dependencies [a9471d0]
+    - @sap-ux/inquirer-common@0.10.2
+    - @sap-ux/project-access@1.32.17
+
+## 0.4.83
+
+### Patch Changes
+
+- @sap-ux/inquirer-common@0.10.1
+
 ## 0.4.82
 
 ### Patch Changes